refactor

+++ src/api/playlists/handler.js
@@ -1,18 +1,16 @@
 const ClientError = require('../../exceptions/ClientError');
 
-class SongsHandler {
+class PlaylistsHandler {
   constructor(service, validator) {
     this._service = service;
     this._validator = validator;
 
     this.postPlaylistHandler = this.postPlaylistHandler.bind(this);
     this.getPlaylistsHandler = this.getPlaylistsHandler.bind(this);
+    this.deletePlaylistByIdHandler = this.deletePlaylistByIdHandler.bind(this);
     this.getSongsInPlaylistHandler = this.getSongsInPlaylistHandler.bind(this);
     this.deleteSongInPlaylistHandler = this.deleteSongInPlaylistHandler.bind(this);
-    // this.getSongByIdHandler = this.getSongByIdHandler.bind(this);
-    // this.putSongByIdHandler = this.putSongByIdHandler.bind(this);
-    this.addSongToPlaylistHandler = this.addSongToPlaylistHandler.bind(this);
-    this.deletePlaylistByIdHandler = this.deletePlaylistByIdHandler.bind(this);
+    this.postSongToPlaylistHandler = this.postSongToPlaylistHandler.bind(this);
   }
 
   async postPlaylistHandler(request, h) {
@@ -70,11 +68,82 @@ class SongsHandler {
     };
   }
 
+  async deletePlaylistByIdHandler(request, h) {
+    try {
+      const { playlistId } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.deletePlaylistById({ playlistId, credentialId });
+
+      return {
+        status: 'success',
+        message: 'Playlist berhasil dihapus',
+      };
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message,
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
+      // Server ERROR!
+      const response = h.response({
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
+      });
+      response.code(500);
+      console.error(error);
+      return response;
+    }
+  }
+
+  async postSongToPlaylistHandler(request, h) {
+    try {
+      this._validator.validatePostSongToPlaylistPayload(request.payload);
+      const { playlistId } = request.params;
+      const { songId } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyPlaylistAccess({ playlistId, credentialId });
+      await this._service.postSongToPlaylist({ songId, playlistId, credentialId });
+
+      const response = h.response({
+        status: 'success',
+        message: 'Lagu berhasil ditambahkan ke playlist',
+      });
+
+      response.code(201);
+      return response;
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message,
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
+      // Server ERROR!
+      const response = h.response({
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
+      });
+      response.code(500);
+      console.error(error);
+      return response;
+    }
+  }
+
   async getSongsInPlaylistHandler(request, h) {
     try {
       const { id: credentialId } = request.auth.credentials;
       const { playlistId } = request.params;
 
+      await this._service.verifyPlaylistAccess({ playlistId, credentialId });
       const songs = await this._service.getSongsInPlaylist({ playlistId, credentialId });
 
       return {
@@ -114,6 +183,7 @@ class SongsHandler {
       const { songId } = request.payload;
       const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifyPlaylistAccess({ playlistId, credentialId });
       await this._service.deleteSongInPlaylist({ playlistId, songId, credentialId });
 
       return {
@@ -140,75 +210,6 @@ class SongsHandler {
       return response;
     }
   }
-
-  async addSongToPlaylistHandler(request, h) {
-    try {
-      this._validator.validatePostAddSongToPlaylist(request.payload);
-      const { playlistId } = request.params;
-      const { songId } = request.payload;
-      const { id: credentialId } = request.auth.credentials;
-
-      await this._service.addSongToPlaylist({ songId, playlistId, credentialId });
-
-      const response = h.response({
-        status: 'success',
-        message: 'Lagu berhasil ditambahkan ke playlist',
-      });
-
-      response.code(201);
-      return response;
-    } catch (error) {
-      if (error instanceof ClientError) {
-        const response = h.response({
-          status: 'fail',
-          message: error.message,
-        });
-        response.code(error.statusCode);
-        return response;
-      }
-
-      // Server ERROR!
-      const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
-      });
-      response.code(500);
-      console.error(error);
-      return response;
-    }
-  }
-
-  async deletePlaylistByIdHandler(request, h) {
-    try {
-      const { playlistId } = request.params;
-      const { id: credentialId } = request.auth.credentials;
-
-      await this._service.deletePlaylistById({ playlistId, credentialId });
-
-      return {
-        status: 'success',
-        message: 'Playlist berhasil dihapus',
-      };
-    } catch (error) {
-      if (error instanceof ClientError) {
-        const response = h.response({
-          status: 'fail',
-          message: error.message,
-        });
-        response.code(error.statusCode);
-        return response;
-      }
-
-      // Server ERROR!
-      const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
-      });
-      response.code(500);
-      console.error(error);
-      return response;
-    }
-  }
 }
 
-module.exports = SongsHandler;
+module.exports = PlaylistsHandler;
